services/HashGenerator/src/main/java/com/pastebin/service/hash/service/impl/HashServiceImpl.java
package com.pastebin.service.hash.service.impl;

import com.pastebin.service.hash.service.HashService;
import org.springframework.stereotype.Service;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.Base64;

@Service
public class HashServiceImpl implements HashService {
    @Override
    public String generateHash(String data) {
        try {
            MessageDigest digest = MessageDigest.getInstance("SHA-256");
            byte[] hash = digest.digest(data.getBytes());
            return Base64.getUrlEncoder().withoutPadding().encodeToString(hash);
        } catch (NoSuchAlgorithmException e) {
            throw new RuntimeException("Failed to generate hash", e);
        }
    }
}